set up conversations/messages

+++ app/controllers/api/v1/comments_controller.rb
@@ -1,29 +1,46 @@
 class Api::V1::CommentsController < Api::V1::BaseController
+  before_action :get_current_user, only: [:create, :update, :destroy]
   before_action :get_comment, only: [:update, :destroy]
 
   def create
-    @comment = Comment.new(comment_params)
+    if @current_user && @current_user.id.to_i == comment_params[:user_id].to_i
+      @comment = Comment.new(comment_params)
 
-    if @comment.save
-      render json: @comment
+      if @comment.save
+        render json: @comment
+      else
+        render json: { errors: @comment.errors.full_messages.to_sentence }, status: 422
+      end
     else
-      render json: { errors: @comment.errors.full_messages.to_sentence }, status: 422
+      render json: { errors: "No current user found" }, status: 422
     end
   end
 
   def update
-    if @comment.update(comment_params)
-      render json: @comment
+    @comment = Comment.find(params[:id])
+
+    if @current_user && @current_user == @comment.user
+      if @comment.update(comment_params)
+        render json: @comment
+      else
+        render json: { errors: @comment.errors.full_messages.to_sentence }, status: 422
+      end
     else
-      render json: { errors: @comment.errors.full_messages.to_sentence }, status: 422
+      render json: { errors: "No current user found" }, status: 422
     end
   end
 
   def destroy
-    if @comment.destroy
-      render json: { success: 'Success' }
+    @comment = Comment.find(params[:id])
+
+    if @current_user && @current_user == @comment.user
+      if @comment.destroy
+        render json: { success: 'Success' }
+      else
+        render json: { errors: @comment.errors.full_messages.to_sentence }, status: 422
+      end
     else
-      render json: { errors: @comment.errors.full_messages.to_sentence }, status: 422
+      render json: { errors: "No current user found" }, status: 422
     end
   end
