+++ app/controllers/api/v1/posts_controller.rb
@@ -1,4 +1,5 @@
 class Api::V1::PostsController < Api::V1::BaseController
+  before_action :get_current_user, only: [:create, :update, :destroy]
   before_action :get_post, only: [:update, :destroy]
 
   def index
@@ -8,28 +9,44 @@ class Api::V1::PostsController < Api::V1::BaseController
   end
 
   def create
-    @post = Post.new(post_params)
+    if @current_user && @current_user.id.to_i == post_params[:user_id].to_i
+      @post = Post.new(post_params)
 
-    if @post.save
-      render json: @post
+      if @post.save
+        render json: @post
+      else
+        render json: { errors: @post.errors.full_messages.to_sentence }, status: 422
+      end
     else
-      render json: { errors: @post.errors.full_messages.to_sentence }, status: 422
+      render json: { errors: "No current user found" }, status: 422
     end
   end
 
   def update
-    if @post.update(post_params)
-      render json: @post
+    @post = Post.find(params[:id])
+
+    if @current_user && @current_user == @post.user
+      if @post.update(post_params)
+        render json: @post
+      else
+        render json: { errors: @post.errors.full_messages.to_sentence }, status: 422
+      end
     else
-      render json: { errors: @post.errors.full_messages.to_sentence }, status: 422
+      render json: { errors: "No current user found" }, status: 422
     end
   end
 
   def destroy
-    if @post.destroy
-      render json: { success: 'Success' }
+    @post = Post.find(params[:id])
+
+    if @current_user && @current_user == @post.user
+      if @post.destroy
+        render json: { success: 'Success' }
+      else
+        render json: { errors: @post.errors.full_messages.to_sentence }, status: 422
+      end
     else
-      render json: { errors: @post.errors.full_messages.to_sentence }, status: 422
+      render json: { errors: "No current user found" }, status: 422
     end
   end
